findempro/requirements/base.txt
# Core Django
Django==4.2.11
djangorestframework==3.15.1
django-environ==0.11.2
python-dotenv==1.0.1

# Database
mysqlclient==2.2.4
redis==5.0.3
hiredis==2.3.2

# Authentication & Security
django-allauth==0.61.1
social-auth-app-django==5.4.0
django-crispy-forms==2.1
django-axes==6.3.0
django-cors-headers==4.3.1
PyJWT==2.8.0

# API Documentation
drf-yasg==1.21.7

# Data Processing
numpy==1.26.4
pandas==2.2.2
scipy==1.13.0
scikit-learn==1.5.0
statsmodels==0.14.1

# Visualization
matplotlib==3.8.4
seaborn==0.13.2

# Simulation
simpy==4.1.1

# Task Queue
celery==5.3.6
django-celery-beat==2.5.0
kombu==5.3.5

# File Processing
Pillow==10.3.0
openpyxl==3.1.2
reportlab==4.1.0

# Monitoring
sentry-sdk==1.45.0
django-prometheus==2.3.1
python-json-logger==2.0.7

# Additional Utilities
django-extensions==3.2.3
django-filter==24.2
django-model-utils==4.3.1
django-simple-history==3.5.0
django-multiselectfield==0.1.13